refactored encoding of 'tild' images
