+++ libheif/codecs/tild.h
@@ -18,8 +18,9 @@
  * along with libheif.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef LIBHEIF_IMAGEITEM_GRID_H
-#define LIBHEIF_IMAGEITEM_GRID_H
+#ifndef LIBHEIF_TILD_H
+#define LIBHEIF_TILD_H
+
 
 #include <codecs/image_item.h>
 #include <vector>
@@ -27,89 +28,101 @@
 #include <memory>
 
 
-class ImageGrid
+#define TILD_OFFSET_NOT_AVAILABLE 0
+#define TILD_OFFSET_SEE_LOWER_RESOLUTION_LAYER 1
+#define TILD_OFFSET_NOT_LOADED 10
+
+class TildHeader
 {
 public:
-  Error parse(const std::vector<uint8_t>& data);
+  void set_parameters(const heif_tild_image_parameters& params);
 
-  std::vector<uint8_t> write() const;
+  Error parse(size_t num_images, const std::vector<uint8_t>& data);
+
+  std::vector<uint8_t> write();
 
   std::string dump() const;
 
-  uint32_t get_width() const { return m_output_width; }
+  uint64_t number_of_tiles() const;
 
-  uint32_t get_height() const { return m_output_height; }
+  uint64_t nTiles_h() const;
 
-  uint16_t get_rows() const
-  {
-    return m_rows;
-  }
+  void set_tild_tile_range(uint32_t tile_x, uint32_t tile_y, uint64_t offset, uint32_t size);
 
-  uint16_t get_columns() const
-  {
-    return m_columns;
-  }
+  size_t get_header_size() const;
 
-  void set_num_tiles(uint16_t columns, uint16_t rows)
-  {
-    m_rows = rows;
-    m_columns = columns;
-  }
+  uint64_t get_tile_offset(uint32_t idx) { return m_offsets[idx].offset; }
 
-  void set_output_size(uint32_t width, uint32_t height)
-  {
-    m_output_width = width;
-    m_output_height = height;
-  }
+  uint32_t get_tile_size(uint32_t idx) { return m_offsets[idx].size; }
 
 private:
-  uint16_t m_rows = 0;
-  uint16_t m_columns = 0;
-  uint32_t m_output_width = 0;
-  uint32_t m_output_height = 0;
+  /*
+   * Flags:
+   * bit 0-1 - number of bits for offsets (0: 32, 1: 40, 2: 48, 3: 64)
+   * bit 2   - with tile sizes
+   * bit 3   - number of bits for size (0: 32, 1: 64)
+   * bit 4   - sequential ordering hint
+   * bit 5   - multidimensional (> 2D)
+   * bit 6   - use 64 bit dimensions (currently unused because ispe is limited to 32 bit)
+   */
+  heif_tild_image_parameters m_parameters;
+
+  struct TileOffset {
+    uint64_t offset = TILD_OFFSET_NOT_LOADED;
+    uint32_t size = 0;
+  };
+
+  std::vector<TileOffset> m_offsets;
+
+  size_t m_header_size = 0;
 };
 
 
-
-
-
-class ImageItem_Grid : public ImageItem
+class ImageItem_Tild : public ImageItem
 {
 public:
-  ImageItem_Grid(HeifContext* ctx, heif_item_id id);
+  ImageItem_Tild(HeifContext* ctx, heif_item_id id);
 
-  ImageItem_Grid(HeifContext* ctx);
+  ImageItem_Tild(HeifContext* ctx);
 
-  const char* get_infe_type() const override { return "grid"; }
+  const char* get_infe_type() const override { return "tild"; }
 
   // const heif_color_profile_nclx* get_forced_output_nclx() const override { return nullptr; }
 
   // heif_compression_format get_compression_format() const override { return heif_compression_HEVC; }
 
+  static Result<std::shared_ptr<ImageItem_Tild>> add_new_tild_item(HeifContext* ctx, const heif_tild_image_parameters* parameters);
+
   Error on_load_file() override;
+
+  void process_before_write() override;
 
   Result<CodedImageData> encode(const std::shared_ptr<HeifPixelImage>& image,
                                 struct heif_encoder* encoder,
                                 const struct heif_encoding_options& options,
                                 enum heif_image_input_class input_class) override {
     return Error{heif_error_Unsupported_feature,
-                 heif_suberror_Unspecified, "Cannot encode image to 'grid'"};
+                 heif_suberror_Unspecified, "Cannot encode image to 'tild'"};
   }
 
   Result<std::shared_ptr<HeifPixelImage>> decode_compressed_image(const struct heif_decoding_options& options,
                                                                   bool decode_tile_only, uint32_t tile_x0, uint32_t tile_y0) const override;
 
 
-  // --- grid specific
+  // --- tild
 
-  const ImageGrid& get_grid_spec() const { return m_grid_spec; }
+  void set_tild_header(const TildHeader& header) { m_tild_header = header; }
 
-  const std::vector<heif_item_id>& get_grid_tiles() const { return m_grid_tile_ids; }
+  TildHeader& get_tild_header() { return m_tild_header; }
+
+  uint64_t get_next_tild_position() { return m_next_tild_position; }
+
+  void set_next_tild_position(uint64_t pos) { m_next_tild_position = pos; }
+
 
 private:
-  ImageGrid m_grid_spec;
-  std::vector<heif_item_id> m_grid_tile_ids;
-
+  TildHeader m_tild_header;
+  uint64_t m_next_tild_position = 0;
 
   Error read_grid_spec();
 
@@ -121,4 +134,4 @@ private:
 };
 
 
-#endif //LIBHEIF_GRID_H
+#endif //LIBHEIF_TILD_H
